requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
paho-mqtt>=2.1,<3.0
